+++ src/reclaim/Claims.sol
@@ -9,7 +9,9 @@ import "./StringUtils.sol";
  * serialising & verifying credentials
  */
 library Claims {
-    /** Data required to describe a claim */
+    /**
+     * Data required to describe a claim
+     */
     struct CompleteClaimData {
         bytes32 identifier;
         address owner;
@@ -23,7 +25,9 @@ library Claims {
         string context;
     }
 
-    /** Claim with signatures & signer */
+    /**
+     * Claim with signatures & signer
+     */
     struct SignedClaim {
         CompleteClaimData claim;
         bytes[] signatures;
@@ -32,10 +36,7 @@ library Claims {
     /**
      * Asserts that the claim is signed by the expected witnesses
      */
-    function assertValidSignedClaim(
-        SignedClaim memory self,
-        address[] memory expectedWitnessAddresses
-    ) internal pure {
+    function assertValidSignedClaim(SignedClaim memory self, address[] memory expectedWitnessAddresses) internal pure {
         require(self.signatures.length > 0, "No signatures");
         address[] memory signedWitnesses = recoverSignersOfSignedClaim(self);
         for (uint256 i = 0; i < expectedWitnessAddresses.length; i++) {
@@ -53,9 +54,7 @@ library Claims {
     /**
      * @dev recovers the signer of the claim
      */
-    function recoverSignersOfSignedClaim(
-        SignedClaim memory self
-    ) internal pure returns (address[] memory) {
+    function recoverSignersOfSignedClaim(SignedClaim memory self) internal pure returns (address[] memory) {
         bytes memory serialised = serialise(self.claim);
         address[] memory signers = new address[](self.signatures.length);
         for (uint256 i = 0; i < self.signatures.length; i++) {
@@ -71,73 +70,49 @@ library Claims {
      *
      * the serialisation is the same as done by the TS library
      */
-    function serialise(
-        CompleteClaimData memory self
-    ) internal pure returns (bytes memory) {
-        return
-            abi.encodePacked(
-                StringUtils.bytes2str(abi.encodePacked(self.identifier)),
-                "\n",
-                StringUtils.address2str(self.owner),
-                "\n",
-                StringUtils.uint2str(self.timestampS),
-                "\n",
-                StringUtils.uint2str(self.epoch)
-            );
+    function serialise(CompleteClaimData memory self) internal pure returns (bytes memory) {
+        return abi.encodePacked(
+            StringUtils.bytes2str(abi.encodePacked(self.identifier)),
+            "\n",
+            StringUtils.address2str(self.owner),
+            "\n",
+            StringUtils.uint2str(self.timestampS),
+            "\n",
+            StringUtils.uint2str(self.epoch)
+        );
     }
 
     /**
      * @dev returns the address of the user that generated the signature
      */
-    function verifySignature(
-        bytes memory content,
-        bytes memory signature
-    ) internal pure returns (address signer) {
-        bytes32 signedHash = keccak256(
-            abi.encodePacked(
-                "\x19Ethereum Signed Message:\n",
-                StringUtils.uint2str(content.length),
-                content
-            )
-        );
+    function verifySignature(bytes memory content, bytes memory signature) internal pure returns (address signer) {
+        bytes32 signedHash =
+            keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n", StringUtils.uint2str(content.length), content));
         return ECDSA.recover(signedHash, signature);
     }
 
-    function hashClaimInfo(
-        ClaimInfo memory claimInfo
-    ) internal pure returns (bytes32) {
-        bytes memory serialised = abi.encodePacked(
-            claimInfo.provider,
-            "\n",
-            claimInfo.parameters,
-            "\n",
-            claimInfo.context
-        );
+    function hashClaimInfo(ClaimInfo memory claimInfo) internal pure returns (bytes32) {
+        bytes memory serialised =
+            abi.encodePacked(claimInfo.provider, "\n", claimInfo.parameters, "\n", claimInfo.context);
         return keccak256(serialised);
     }
 
     /**
      * @dev performs string matching to find fields in context
      */
-    function extractFieldFromContext(
-        string memory data,
-        string memory target
-    ) public pure returns (string memory) {
+    function extractFieldFromContext(string memory data, string memory target) public pure returns (string memory) {
         bytes memory dataBytes = bytes(data);
         bytes memory targetBytes = bytes(target);
 
-        require(
-            dataBytes.length >= targetBytes.length,
-            "target is longer than data"
-        );
-        uint start = 0;
+        require(dataBytes.length >= targetBytes.length, "target is longer than data");
+        uint256 start = 0;
         bool foundStart = false;
         // Find start of "contextMessage":"
 
-        for (uint i = 0; i <= dataBytes.length - targetBytes.length; i++) {
+        for (uint256 i = 0; i <= dataBytes.length - targetBytes.length; i++) {
             bool isMatch = true;
 
-            for (uint j = 0; j < targetBytes.length && isMatch; j++) {
+            for (uint256 j = 0; j < targetBytes.length && isMatch; j++) {
                 if (dataBytes[i + j] != targetBytes[j]) {
                     isMatch = false;
                 }
@@ -158,24 +133,18 @@ library Claims {
         // The function does not need to handle escaped backslashes specifically because
         // it only looks for the first unescaped quote to mark the end of the field value.
         // Escaped quotes (preceded by a backslash) are naturally ignored in this logic.
-        uint end = start;
-        while (
-            end < dataBytes.length &&
-            !(dataBytes[end] == '"' && dataBytes[end - 1] != "\\")
-        ) {
+        uint256 end = start;
+        while (end < dataBytes.length && !(dataBytes[end] == '"' && dataBytes[end - 1] != "\\")) {
             end++;
         }
 
         // if the end is not found, return an empty string because of malformed or missing message
-        if (
-            end <= start ||
-            !(dataBytes[end] == '"' && dataBytes[end - 1] != "\\")
-        ) {
+        if (end <= start || !(dataBytes[end] == '"' && dataBytes[end - 1] != "\\")) {
             return ""; // Malformed or missing message
         }
 
         bytes memory contextMessage = new bytes(end - start);
-        for (uint i = start; i < end; i++) {
+        for (uint256 i = start; i < end; i++) {
             contextMessage[i - start] = dataBytes[i];
         }
         return string(contextMessage);
