+++ src/reclaim/Reclaim.sol
@@ -13,25 +13,38 @@ import "./BytesUtils.sol";
  */
 contract Reclaim {
     struct Witness {
-        /** ETH address of the witness */
+        /**
+         * ETH address of the witness
+         */
         address addr;
-        /** Host to connect to the witness */
+        /**
+         * Host to connect to the witness
+         */
         string host;
     }
 
     struct Epoch {
-        /** Epoch number */
+        /**
+         * Epoch number
+         */
         uint32 id;
-        /** when the epoch changed */
+        /**
+         * when the epoch changed
+         */
         uint32 timestampStart;
-        /** when the epoch will change */
+        /**
+         * when the epoch will change
+         */
         uint32 timestampEnd;
-        /** Witnesses for this epoch */
+        /**
+         * Witnesses for this epoch
+         */
         Witness[] witnesses;
         /**
          * Minimum number of witnesses
          * required to create a claim
-         * */
+         *
+         */
         uint8 minimumWitnessesForClaimCreation;
     }
 
@@ -40,20 +53,24 @@ contract Reclaim {
         Claims.SignedClaim signedClaim;
     }
 
-    /** list of all epochs */
+    /**
+     * list of all epochs
+     */
     Epoch[] public epochs;
 
     /**
      * duration of each epoch.
      * is not a hard duration, but useful for
      * caching purposes
-     * */
+     *
+     */
     uint32 public epochDurationS; // 1 day
 
     /**
      * current epoch.
      * starts at 1, so that the first epoch is 1
-     * */
+     *
+     */
     uint32 public currentEpoch;
 
     event EpochAdded(Epoch epoch);
@@ -92,11 +109,11 @@ contract Reclaim {
     /**
      * Get the witnesses that'll sign the claim
      */
-    function fetchWitnessesForClaim(
-        uint32 epoch,
-        bytes32 identifier,
-        uint32 timestampS
-    ) public view returns (Witness[] memory) {
+    function fetchWitnessesForClaim(uint32 epoch, bytes32 identifier, uint32 timestampS)
+        public
+        view
+        returns (Witness[] memory)
+    {
         Epoch memory epochData = fetchEpoch(epoch);
         bytes memory completeInput = abi.encodePacked(
             // hex encode bytes
@@ -114,19 +131,13 @@ contract Reclaim {
         bytes memory completeHash = abi.encodePacked(keccak256(completeInput));
 
         Witness[] memory witnessesLeftList = epochData.witnesses;
-        Witness[] memory selectedWitnesses = new Witness[](
-            epochData.minimumWitnessesForClaimCreation
-        );
-        uint witnessesLeft = witnessesLeftList.length;
+        Witness[] memory selectedWitnesses = new Witness[](epochData.minimumWitnessesForClaimCreation);
+        uint256 witnessesLeft = witnessesLeftList.length;
 
-        uint byteOffset = 0;
-        for (
-            uint32 i = 0;
-            i < epochData.minimumWitnessesForClaimCreation;
-            i++
-        ) {
-            uint randomSeed = BytesUtils.bytesToUInt(completeHash, byteOffset);
-            uint witnessIndex = randomSeed % witnessesLeft;
+        uint256 byteOffset = 0;
+        for (uint32 i = 0; i < epochData.minimumWitnessesForClaimCreation; i++) {
+            uint256 randomSeed = BytesUtils.bytesToUInt(completeHash, byteOffset);
+            uint256 witnessIndex = randomSeed % witnessesLeft;
             selectedWitnesses[i] = witnessesLeftList[witnessIndex];
             // remove the witness from the list of witnesses
             // we've utilised witness at index "idx"
@@ -135,9 +146,7 @@ contract Reclaim {
             // and reduce the number of witnesses left to pick from
             // since solidity doesn't support "pop()" in memory arrays
             // we swap the last element with the element we want to remove
-            witnessesLeftList[witnessIndex] = epochData.witnesses[
-                witnessesLeft - 1
-            ];
+            witnessesLeftList[witnessIndex] = epochData.witnesses[witnessesLeft - 1];
             byteOffset = (byteOffset + 4) % completeHash.length;
             witnessesLeft -= 1;
         }
@@ -152,10 +161,7 @@ contract Reclaim {
     function verifyProof(Proof memory proof) public view {
         // create signed claim using claimData and signature.
         require(proof.signedClaim.signatures.length > 0, "No signatures");
-        Claims.SignedClaim memory signed = Claims.SignedClaim(
-            proof.signedClaim.claim,
-            proof.signedClaim.signatures
-        );
+        Claims.SignedClaim memory signed = Claims.SignedClaim(proof.signedClaim.claim, proof.signedClaim.signatures);
 
         // check if the hash from the claimInfo is equal to the infoHash in the claimData
         bytes32 hashed = Claims.hashClaimInfo(proof.claimInfo);
@@ -163,23 +169,18 @@ contract Reclaim {
 
         // fetch witness list from fetchEpoch(_epoch).witnesses
         Witness[] memory expectedWitnesses = fetchWitnessesForClaim(
-            proof.signedClaim.claim.epoch,
-            proof.signedClaim.claim.identifier,
-            proof.signedClaim.claim.timestampS
+            proof.signedClaim.claim.epoch, proof.signedClaim.claim.identifier, proof.signedClaim.claim.timestampS
         );
-        address[] memory signedWitnesses = Claims.recoverSignersOfSignedClaim(
-            signed
-        );
+        address[] memory signedWitnesses = Claims.recoverSignersOfSignedClaim(signed);
         // check if the number of signatures is equal to the number of witnesses
         require(
-            signedWitnesses.length == expectedWitnesses.length,
-            "Number of signatures not equal to number of witnesses"
+            signedWitnesses.length == expectedWitnesses.length, "Number of signatures not equal to number of witnesses"
         );
 
         // Update awaited: more checks on whose signatures can be considered.
         for (uint256 i = 0; i < signed.signatures.length; i++) {
             bool found = false;
-            for (uint j = 0; j < expectedWitnesses.length; j++) {
+            for (uint256 j = 0; j < expectedWitnesses.length; j++) {
                 if (signedWitnesses[i] == expectedWitnesses[j].addr) {
                     found = true;
                     break;
@@ -196,10 +197,7 @@ contract Reclaim {
     /**
      * @dev Add a new epoch
      */
-    function addNewEpoch(
-        Witness[] calldata witnesses,
-        uint8 requisiteWitnessesForClaimCreate
-    ) external onlyOwner {
+    function addNewEpoch(Witness[] calldata witnesses, uint8 requisiteWitnessesForClaimCreate) external onlyOwner {
         if (epochDurationS == 0) {
             epochDurationS = 1 days;
         }
@@ -212,8 +210,7 @@ contract Reclaim {
         epoch.id = currentEpoch;
         epoch.timestampStart = uint32(block.timestamp);
         epoch.timestampEnd = uint32(block.timestamp + epochDurationS);
-        epoch
-            .minimumWitnessesForClaimCreation = requisiteWitnessesForClaimCreate;
+        epoch.minimumWitnessesForClaimCreation = requisiteWitnessesForClaimCreate;
 
         for (uint256 i = 0; i < witnesses.length; i++) {
             epoch.witnesses.push(witnesses[i]);
@@ -224,10 +221,7 @@ contract Reclaim {
 
     // internal code -----
 
-    function uintDifference(
-        uint256 a,
-        uint256 b
-    ) internal pure returns (uint256) {
+    function uintDifference(uint256 a, uint256 b) internal pure returns (uint256) {
         if (a > b) {
             return a - b;
         }
